Split sync into submodules
